use default implementation for set_column
