+++ src/integer/mat_poly_over_z/set.rs
@@ -114,67 +114,6 @@ impl MatrixSetEntry<&PolyOverZ> for MatPolyOverZ {
 implement_for_owned!(PolyOverZ, MatPolyOverZ, MatrixSetEntry);
 
 impl MatrixSetSubmatrix for MatPolyOverZ {
-    /// Sets a column of the given matrix to the provided column of `other`.
-    ///
-    /// Parameters:
-    /// - `col_0`: specifies the column of `self` that should be modified
-    /// - `other`: specifies the matrix providing the column replacing the column in `self`
-    /// - `col_1`: specifies the column of `other` providing
-    ///   the values replacing the original column in `self`
-    ///
-    /// Negative indices can be used to index from the back, e.g., `-1` for
-    /// the last element.
-    ///
-    /// Returns an empty `Ok` if the action could be performed successfully.
-    /// Otherwise, a [`MathError`] is returned if one of the specified columns is not part of its matrix
-    /// or if the number of rows differs.
-    ///
-    /// # Examples
-    /// ```
-    /// use qfall_math::{integer::MatPolyOverZ, traits::MatrixSetSubmatrix};
-    /// use std::str::FromStr;
-    ///
-    /// let mut mat_1 = MatPolyOverZ::new(2, 2);
-    /// let mat_2 = MatPolyOverZ::from_str("[[1  1],[0]]").unwrap();
-    /// mat_1.set_column(1, &mat_2, 0);
-    /// ```
-    ///
-    /// # Errors and Failures
-    /// - Returns a [`MathError`] of type [`MathError::OutOfBounds`]
-    ///   if the provided column index is not defined within the margins of the matrix.
-    /// - Returns a [`MathError`] of type [`MismatchingMatrixDimension`](MathError::MismatchingMatrixDimension)
-    ///   if the number of rows of `self` and `other` differ.
-    fn set_column(
-        &mut self,
-        col_0: impl TryInto<i64> + Display,
-        other: &Self,
-        col_1: impl TryInto<i64> + Display,
-    ) -> Result<(), MathError> {
-        let num_rows_0 = self.get_num_rows();
-        let num_rows_1 = other.get_num_rows();
-        let num_cols_0 = self.get_num_columns();
-        let num_cols_1 = other.get_num_columns();
-        let col_0 = evaluate_index_for_vector(col_0, num_cols_0)?;
-        let col_1 = evaluate_index_for_vector(col_1, num_cols_1)?;
-
-        if num_rows_0 != num_rows_1 {
-            return Err(MathError::MismatchingMatrixDimension(format!(
-                "as set_column was called on two matrices with different number of rows/columns {num_rows_0} and {num_rows_1}",
-            )));
-        }
-
-        for row in 0..num_rows_0 {
-            unsafe {
-                fmpz_poly_set(
-                    fmpz_poly_mat_entry(&self.matrix, row, col_0),
-                    fmpz_poly_mat_entry(&other.matrix, row, col_1),
-                )
-            };
-        }
-
-        Ok(())
-    }
-
     unsafe fn set_submatrix_unchecked(
         &mut self,
         row_self_start: i64,
