+++ src/integer_mod_q/mat_polynomial_ring_zq/set.rs
@@ -11,7 +11,7 @@
 use super::MatPolynomialRingZq;
 use crate::integer_mod_q::PolynomialRingZq;
 use crate::macros::for_others::implement_for_owned;
-use crate::traits::{CompareBase, MatrixSetSubmatrix, MatrixSwaps};
+use crate::traits::{MatrixSetSubmatrix, MatrixSwaps};
 use crate::utils::index::evaluate_indices_for_matrix;
 use crate::{error::MathError, integer::PolyOverZ, traits::MatrixSetEntry};
 use flint_sys::fmpz_poly_mat::{
@@ -216,51 +216,6 @@ implement_for_owned!(PolyOverZ, MatPolynomialRingZq, MatrixSetEntry);
 implement_for_owned!(PolynomialRingZq, MatPolynomialRingZq, MatrixSetEntry);
 
 impl MatrixSetSubmatrix for MatPolynomialRingZq {
-    /// Sets a column of the given matrix to the provided column of `other`.
-    ///
-    /// Parameters:
-    /// - `col_0`: specifies the column of `self` that should be modified
-    /// - `other`: specifies the matrix providing the column replacing the column in `self`
-    /// - `col_1`: specifies the column of `other` providing
-    ///   the values replacing the original column in `self`
-    ///
-    /// Negative indices can be used to index from the back, e.g., `-1` for
-    /// the last element.
-    ///
-    /// Returns an empty `Ok` if the action could be performed successfully.
-    /// Otherwise, a [`MathError`] is returned if one of the specified columns is not part of its matrix
-    /// or if the number of rows differs.
-    ///
-    /// # Examples
-    /// ```
-    /// use qfall_math::{integer_mod_q::MatPolynomialRingZq, traits::MatrixSetSubmatrix};
-    /// use std::str::FromStr;
-    ///
-    /// let mut mat_1 = MatPolynomialRingZq::from_str("[[0, 0, 0],[0, 0, 0]] / 2  1 1 mod 6").unwrap();
-    /// let mat_2 = MatPolynomialRingZq::from_str("[[1  2, 0, 1  3],[1  3, 1  4, 1  5]] / 2  1 1 mod 6").unwrap();
-    /// mat_1.set_column(1, &mat_2, 0);
-    /// ```
-    ///
-    /// # Errors and Failures
-    /// - Returns a [`MathError`] of type [`MathError::OutOfBounds`]
-    ///   if the provided column index is not defined within the margins of the matrix.
-    /// - Returns a [`MathError`] of type [`MismatchingMatrixDimension`](MathError::MismatchingMatrixDimension)
-    ///   if the number of rows of `self` and `other` differ.
-    /// - Returns a [`MathError`] of type [`MismatchingModulus`](MathError::MismatchingModulus)
-    ///   if the moduli of `self` and `other` mismatch.
-    fn set_column(
-        &mut self,
-        col_0: impl TryInto<i64> + Display,
-        other: &Self,
-        col_1: impl TryInto<i64> + Display,
-    ) -> Result<(), MathError> {
-        if !self.compare_base(other) {
-            return Err(self.call_compare_base_error(other).unwrap());
-        }
-
-        self.matrix.set_column(col_0, &other.matrix, col_1)
-    }
-
     unsafe fn set_submatrix_unchecked(
         &mut self,
         row_self_start: i64,
